requirement.txt
fastapi
pydantic 
google-generativeai
pymongo
python-dotenv
uvicorn
Pillow